update decimal prices for integer prices

+++ MarketOrganizer.Data/Models/Item.cs
@@ -12,8 +12,8 @@ namespace MarketOrganizer.Data.Models
   {
     public int Id { get; set; }
     public string ItemName { get; set; }
-    public decimal BuyPrice { get; set; }
-    public decimal SellPrice { get; set; }
-    public decimal HighestMarketPrice { get; set; }
+    public int BuyPrice { get; set; }
+    public int SellPrice { get; set; }
+    public int HighestMarketPrice { get; set; }
   }
 }
